Fix Empty adSlot Object Caused by Ad Blocker
Issue

When using an ad blocker adSlot can become an empty object which
causes the call to adSlot.getServices to produce the error:
"getServices is not a function".

Solution

By adding the check adSlot.hasOwnProperty("getServices") we can ensure
that the adSlot object has the correct function before attempting to
call it.