General improvement: Update host for the gpt.js to improve performance
https://developers.google.com/publisher-ads-audits/reference/audits/loads-gpt-from-sgdn
```
"The Google Publisher Tag library is now hosted at https://securepubads.g.doubleclick.net/tag/js/gpt.js, in addition to being hosted at the googletagservices.com domain. While not required, we strongly recommend that you update all references to GPT on your pages to use this new domain.
This change consolidates all ad serving requests to one domain, instead of two, which means the browser only needs to connect to one domain. The library is exactly the same at both domains.
The result is an improvement in the speed of your tags and fetching ads a bit quicker."
```